Extract remaining parts out of ApproveOrderFragment into CardViewModel.
