+++ Demo/src/main/java/com/paypal/android/ui/card/CardViewModel.kt
@@ -1,17 +1,133 @@
 package com.paypal.android.ui.card
 
+import android.content.Context
+import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.paypal.android.api.model.Order
 import com.paypal.android.api.model.OrderIntent
+import com.paypal.android.api.services.SDKSampleServerAPI
+import com.paypal.android.cardpayments.ApproveOrderListener
+import com.paypal.android.cardpayments.Card
+import com.paypal.android.cardpayments.CardClient
+import com.paypal.android.cardpayments.CardRequest
 import com.paypal.android.cardpayments.CardResult
+import com.paypal.android.cardpayments.threedsecure.SCA
+import com.paypal.android.corepayments.CoreConfig
+import com.paypal.android.corepayments.PayPalSDKError
+import com.paypal.android.fraudprotection.PayPalDataCollector
+import com.paypal.android.models.OrderRequest
 import com.paypal.android.models.TestCard
+import com.paypal.android.usecase.CompleteOrderUseCase
+import com.paypal.android.usecase.CreateOrderUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class CardViewModel : ViewModel() {
+@HiltViewModel
+class CardViewModel @Inject constructor(
+    val createOrderUseCase: CreateOrderUseCase,
+    val sdkSampleServerAPI: SDKSampleServerAPI,
+    val completeOrderUseCase: CompleteOrderUseCase,
+) : ViewModel() {
+
+    companion object {
+        const val TAG = "CardFragment"
+        const val APP_RETURN_URL = "com.paypal.android.demo://example.com/returnUrl"
+    }
+
     private val _uiState = MutableStateFlow(CardViewUiState())
     val uiState = _uiState.asStateFlow()
+
+    private lateinit var cardClient: CardClient
+    private lateinit var payPalDataCollector: PayPalDataCollector
+
+    fun createOrder() {
+        viewModelScope.launch {
+            isCreateOrderLoading = true
+
+            val uiState = uiState.value
+            val orderRequest = uiState.run { OrderRequest(intentOption, shouldVault, customerId) }
+            createdOrder = createOrderUseCase(orderRequest)
+            isCreateOrderLoading = false
+        }
+    }
+
+    fun approveOrder(activity: AppCompatActivity) {
+        viewModelScope.launch {
+            isApproveOrderLoading = true
+
+            val clientId = sdkSampleServerAPI.fetchClientId()
+            val coreConfig = CoreConfig(clientId = clientId)
+            payPalDataCollector = PayPalDataCollector(coreConfig)
+
+            cardClient = CardClient(activity, coreConfig)
+            cardClient.approveOrderListener = object : ApproveOrderListener {
+                override fun onApproveOrderSuccess(result: CardResult) {
+                    approveOrderResult = result
+                    isApproveOrderLoading = false
+                }
+
+                override fun onApproveOrderFailure(error: PayPalSDKError) {
+                    approveOrderErrorMessage = "CAPTURE fail: ${error.errorDescription}"
+                    isApproveOrderLoading = false
+                }
+
+                override fun onApproveOrderCanceled() {
+                    approveOrderErrorMessage = "USER CANCELED"
+                    isApproveOrderLoading = false
+                }
+
+                override fun onApproveOrderThreeDSecureWillLaunch() {
+                    Log.d(TAG, "3DS Auth Requested")
+                }
+
+                override fun onApproveOrderThreeDSecureDidFinish() {
+                    Log.d(TAG, "3DS Success")
+                    isApproveOrderLoading = false
+                }
+            }
+
+            val cardRequest = createCardRequest(uiState.value, createdOrder!!)
+            cardClient.approveOrder(activity, cardRequest)
+        }
+    }
+
+    fun completeOrder(context: Context) {
+        viewModelScope.launch {
+            isCompleteOrderLoading = true
+
+            val cmid = payPalDataCollector.collectDeviceData(context)
+            val orderId = createdOrder!!.id!!
+
+            completedOrder = completeOrderUseCase(orderId, intentOption, cmid)
+            isCompleteOrderLoading = false
+        }
+    }
+
+    private fun createCardRequest(uiState: CardViewUiState, order: Order): CardRequest {
+        val card = parseCard(uiState)
+        val sca = when (uiState.scaOption) {
+            "ALWAYS" -> SCA.SCA_ALWAYS
+            else -> SCA.SCA_WHEN_REQUIRED
+        }
+        return CardRequest(order.id!!, card, APP_RETURN_URL, sca)
+    }
+
+    private fun parseCard(uiState: CardViewUiState): Card {
+        // expiration date in UI State needs to be formatted because it uses a visual transformation
+        val dateString = DateString(uiState.cardExpirationDate)
+        return Card(
+            number = uiState.cardNumber,
+            expirationMonth = dateString.formattedMonth,
+            expirationYear = dateString.formattedYear,
+            securityCode = uiState.cardSecurityCode
+        )
+    }
 
     var createdOrder: Order?
         get() = _uiState.value.createdOrder
